@@ -1,8 +1,8 @@
 package id.duglegir.jagosholat.Controller.ViewPagerAdapter;
 
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
-import android.support.v4.app.FragmentPagerAdapter;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
+import androidx.fragment.app.FragmentPagerAdapter;
 
 import id.duglegir.jagosholat.Controller.MainActivityContent.JadwalFragment;
 import id.duglegir.jagosholat.Controller.MainActivityContent.CatatanFragment;
