@@ -1,7 +1,7 @@
 package id.duglegir.jagosholat.View;
 
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 
 import id.duglegir.jagosholat.R;
 
